requirements.txt
pydantic[email]==2.11.2
httpx==0.28.1
faiss-cpu==1.10.0
langchain-community==0.3.21
fastapi==0.115.12
uvicorn[standard]==0.34.1
asyncpg==0.30
python-dotenv==1.1.0
sqlalchemy==2.0.40
PyJWT==2.9.0
langchain_huggingface==0.1.2
rank_bm25==0.2.2